Update dockerized scripts (Also run linting)
