+++ scripts/compile/Dockerfile
@@ -1,0 +1,6 @@
+FROM mhart/alpine-node
+COPY run .
+RUN chmod +x run
+COPY package.json .
+RUN npm install
+CMD ./run
