+++ scripts/demo/demo/entrypoint.js
@@ -1,9 +1,9 @@
-require('babel/register')({
+require('babel-register')({
   // This will override `node_modules` ignoring - you can alternatively pass
   // an array of strings to be explicitly matched or a regex / glob
-  ignore: false
+  ignore: false,
   //Set the experimental proposal stage.
   //More: (http://babeljs.io/docs/usage/experimental/)
-  ,stage:0
+  presets: ["stage-0", "es2015"],
 });
-require('./test.js');
+require('./index.js');
