+++ scripts/demo/demo/index.js
@@ -1,0 +1,8 @@
+import Understate from './../PROJECT/dist';
+const log   = value => console.log(value);
+var increment = x => x + 1;
+const state     = new Understate({initial: 0});
+state.subscribe(log);
+state.set(increment);//1
+state.set(increment);//2
+state.set(increment);//3
