+++ scripts/demo/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "demo",
+  "version": "0.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "demo": "node demo/entrypoint.js"
+  },
+  "dependencies":{
+    "babel": "^5.8.23",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.14.0"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
