+++ scripts/demo/run
@@ -1,0 +1,2 @@
+#! /bin/sh
+npm run demo -s
