+++ scripts/deploy/Dockerfile
@@ -1,0 +1,5 @@
+FROM mhart/alpine-node
+COPY package.json .
+COPY run .
+RUN chmod +x run
+CMD ./run
