+++ scripts/deploy/run
@@ -1,0 +1,2 @@
+#! /bin/sh
+npm run deploy -s
