+++ scripts/lint/npm-debug.log
@@ -1,0 +1,48 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
+1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'run',
+1 verbose cli   'lint' ]
+2 info using npm@3.8.9
+3 info using node@v6.2.0
+4 verbose run-script [ 'prelint', 'lint', 'postlint' ]
+5 info lifecycle lint@0.0.0~prelint: lint@0.0.0
+6 silly lifecycle lint@0.0.0~prelint: no script for prelint, continuing
+7 info lifecycle lint@0.0.0~lint: lint@0.0.0
+8 verbose lifecycle lint@0.0.0~lint: unsafe-perm in lifecycle true
+9 verbose lifecycle lint@0.0.0~lint: PATH: C:\Program Files\nodejs\node_modules\npm\bin\node-gyp-bin;C:\Users\john\BTSync\projects\understate\scripts\lint\node_modules\.bin;C:\Program Files\nodejs;C:\Program Files\Docker\Docker\Resources\bin;C:\WINDOWS\system32;C:\WINDOWS;C:\WINDOWS\System32\Wbem;C:\WINDOWS\System32\WindowsPowerShell\v1.0\;C:\Program Files\Intel\WiFi\bin\;C:\Program Files\Common Files\Intel\WirelessCommon\;C:\Program Files\nodejs\;C:\WINDOWS\system32\config\systemprofile\.dnx\bin;C:\Program Files\Microsoft DNX\Dnvm\;C:\Program Files\Microsoft SQL Server\130\Tools\Binn\;C:\Program Files\Git\cmd;C:\Program Files\OpenSSH\bin;C:\Users\john\AppData\Local\atom\bin;C:\Program Files (x86)\Microsoft VS Code\bin;C:\Users\john\Application Data\npm\node_modules;C:\Users\john\AppData\Roaming\npm;C:\Program Files (x86)\Nmap;C:\Users\john\AppData\Local\Microsoft\WindowsApps;C:\Program Files\terraform;C:\Users\john\AppData\Local\GitHubCLI\bin
+10 verbose lifecycle lint@0.0.0~lint: CWD: C:\Users\john\BTSync\projects\understate\scripts\lint
+11 silly lifecycle lint@0.0.0~lint: Args: [ '/d /s /c', 'eslint -c .eslintrc.json src/' ]
+12 silly lifecycle lint@0.0.0~lint: Returned: code: 1  signal: null
+13 info lifecycle lint@0.0.0~lint: Failed to exec lint script
+14 verbose stack Error: lint@0.0.0 lint: `eslint -c .eslintrc.json src/`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\lifecycle.js:245:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\spawn.js:24:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:850:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:215:5)
+15 verbose pkgid lint@0.0.0
+16 verbose cwd C:\Users\john\BTSync\projects\understate\scripts\lint
+17 error Windows_NT 10.0.14393
+18 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "run" "lint"
+19 error node v6.2.0
+20 error npm  v3.8.9
+21 error code ELIFECYCLE
+22 error lint@0.0.0 lint: `eslint -c .eslintrc.json src/`
+22 error Exit status 1
+23 error Failed at the lint@0.0.0 lint script 'eslint -c .eslintrc.json src/'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the lint package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     eslint -c .eslintrc.json src/
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs lint
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls lint
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
