+++ scripts/test/Dockerfile
@@ -1,0 +1,7 @@
+FROM mhart/alpine-node
+COPY run .
+RUN chmod +x run
+COPY package.json .
+RUN npm install
+COPY test /test
+CMD ./run
