+++ scripts/test/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "test",
+  "version": "0.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "mocha test/entrypoint.js"
+  },
+  "dependencies": {
+    "babel": "^5.8.23",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.14.0",
+    "mocha": "^2.4.5",
+    "rollup": "^0.26.2"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
