+++ scripts/test/run
@@ -1,0 +1,2 @@
+#! /bin/sh
+npm run test -s
