+++ scripts/test/test/entrypoint.js
@@ -1,0 +1,9 @@
+require('babel-register')({
+  // This will override `node_modules` ignoring - you can alternatively pass
+  // an array of strings to be explicitly matched or a regex / glob
+  ignore: false,
+  //Set the experimental proposal stage.
+  //More: (http://babeljs.io/docs/usage/experimental/)
+  presets: ["stage-0", "es2015"],
+});
+require('./index.js');
