+++ scripts/test/test/notLocalStorage.js
